refactor: domain 패키지 생성
- DTO와 domain을 패키지로 구분
