+++ src/main/java/gift/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.DTO.User;
+import gift.domain.User;
 import java.util.Optional;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
