+++ src/main/java/gift/repository/WishlistRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.DTO.Wishlist;
+import gift.domain.Wishlist;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
